build_agent_tool now a method on the client
